% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{coefficients_lm}
\alias{coefficients_lm}
\title{Coefficients from lm models}
\usage{
coefficients_lm(lst)
}
\arguments{
\item{lst}{List of lm models}
}
\value{
Table with coefficients for each model
}
\description{
Coefficients from lm models
}
\examples{
coefficients.lm(lst)
}
