% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{coefficients_step}
\alias{coefficients_step}
\title{Coefficients from stepwise models}
\usage{
coefficients_step(lst)
}
\arguments{
\item{lst}{List of stepwise models}
}
\value{
Table with coefficients for each model
}
\description{
Coefficients from stepwise models
}
\examples{
coefficients.step(lst)
}
