% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{convertPricesToReturns}
\alias{convertPricesToReturns}
\title{Converts prices to returns}
\usage{
convertPricesToReturns(prices, freq = "D")
}
\arguments{
\item{prices}{xts of prices such as returned by getPrices function}

\item{freq}{M for monthly, or D for daily}
}
\value{
If ncol(prices)==1, an xts object, else a list with xts object of returns
}
\description{
Converts prices to returns
}
\examples{
convertPricesToReturns(p)
}
