% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{dateFilter}
\alias{dateFilter}
\title{Filter (subset) an xts object
Subsets an xts object from start (s) to end (e).  If either is omitted, the earliest or latest observation is used
If n is not null it will return n observations from the last date (e).}
\usage{
dateFilter(xtsData, s = NULL, e = NULL, n = NULL)
}
\arguments{
\item{xtsData}{xts object to be filtered (subsetted)}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of observations}
}
\value{
xts object
}
\description{
Filter (subset) an xts object
Subsets an xts object from start (s) to end (e).  If either is omitted, the earliest or latest observation is used
If n is not null it will return n observations from the last date (e).
}
\examples{
dateFilter(xtsOjbect,s="2012-12-31",e="2017-12-31")
dateFilter(xtsOjbect,e="2017-12-31", n=60)
}
