% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{defineBenchmark}
\alias{defineBenchmark}
\title{Define a benchmark    
A benchmark is used to compare the performance of a fund against. A benchmark may be a single symbol with a weight of 1,
or a blend of multiple symbols (a vector)}
\usage{
defineBenchmark(shortName = "Bench", description = "Benchmark",
  symbol = NULL, weights = 1, startDate = "1970-01-01", freq = "M")
}
\arguments{
\item{shortName}{A short name used much the way a symbol (ticker) might be used as a column head.}

\item{description}{A longer description of the benchmark (e.g., 50% S&P 500 / 50% Barclay's Agg)}

\item{symbol}{A symbol (ticker) or character vector of symbols}

\item{weights}{Weights of the symbol(s). Length must match the symbol}

\item{startDate}{Start date to get the prices for the benchmarks}

\item{freq}{Frequency of returns for the benchmark ("M" or "D")}
}
\value{
List with information about the benchmark
}
\description{
Define a benchmark    
A benchmark is used to compare the performance of a fund against. A benchmark may be a single symbol with a weight of 1,
or a blend of multiple symbols (a vector)
}
\examples{
defineBenchmark("SP500","S&P 500 ETF", "SPY", 1)
defineBenchmark("Bench","50\% S&P 500 / 50\% Barclay's Agg", c("SPY","AGG"), c(0.5,0.5))
}
