% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{download_FF_5_factor_monthly}
\alias{download_FF_5_factor_monthly}
\title{Download Fama-French 5 factor monthly     
source: http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/F-F_Research_Data_5_Factors_2x3_CSV.zip}
\usage{
download_FF_5_factor_monthly()
}
\value{
xts object with monthly data for Fama-French model
}
\description{
Download Fama-French 5 factor monthly     
source: http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/F-F_Research_Data_5_Factors_2x3_CSV.zip
}
\examples{
download_FF_5_factor_monthly()
}
