% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{fa.style.fit}
\alias{fa.style.fit}
\title{Calculate effective style weights
This is a clone of the style.fit function from the Factor Analytics package. The only difference is
that it includes a trace parameter to pass to the step function to control its output.}
\usage{
fa.style.fit(R.fund, R.style, model = FALSE, method = c("constrained",
  "unconstrained", "normalized"), leverage = FALSE, selection = c("none",
  "AIC"), trace = 0, ...)
}
\arguments{
\item{R.fund}{matrix, data frame, or zoo object with fund returns to be analyzed}

\item{R.style}{matrix, data frame, or zoo object with style index returns. Data object must be of the same length and time-aligned with R.fund}

\item{model}{logical. If 'model' = TRUE in style.QPfit, the full result set is shown from the output of solve.QP.}

\item{method}{specify the method of calculation of style weights as "constrained", "unconstrained", or "normalized". For more information, see style.fit}

\item{leverage}{logical, defaults to 'FALSE'. If 'TRUE', the calculation of weights assumes that leverage may be used. For more information, see style.fit}

\item{selection}{either "none" (default) or "AIC". If "AIC", then the function uses a stepwise regression to identify find the model with minimum AIC value. See step for more detail.}

\item{trace}{value passed to the step function to control feedback. Default is zero to suppress printing.}

\item{...}{}
}
\value{
list with weights and r-squared values.
}
\description{
Calculate effective style weights
This is a clone of the style.fit function from the Factor Analytics package. The only difference is
that it includes a trace parameter to pass to the step function to control its output.
}
\examples{
fa.style.fit(R.fund, R.style)
}
