% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{faAlignXTS}
\alias{faAlignXTS}
\title{Align date indices of two xts objects}
\usage{
faAlignXTS(xts1, xts2, s = NULL, e = NULL, n = NULL)
}
\arguments{
\item{xts1}{First xts object}

\item{xts2}{Second xts object}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of periods}
}
\value{
List with two xts objects with the same indices
}
\description{
Align date indices of two xts objects
}
\examples{
faAlignXTS(xts1, xts2)
}
