% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{faCommonDate}
\alias{faCommonDate}
\title{Find common period between xts objects    
Given a list of xts objects, it returns a list of those objects with the longest common period (starting at most recent start,
and ending at earliest end)}
\usage{
faCommonDate(rets)
}
\arguments{
\item{rets}{List of xts objects}
}
\value{
List of xts objects with common dates
}
\description{
Find common period between xts objects    
Given a list of xts objects, it returns a list of those objects with the longest common period (starting at most recent start,
and ending at earliest end)
}
\examples{
faCommonDate(rets)
}
