% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{ffMergeXTS}
\alias{ffMergeXTS}
\title{Merge return series with Fama-French data
Merges a xts object of returns with the ff data xts object.}
\usage{
ffMergeXTS(y, ff_data, s = NULL, e = NULL, n = NULL)
}
\arguments{
\item{y}{Fund return}

\item{ff_data}{Fama-French data}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of observations}
}
\value{
xts object
}
\description{
Merge return series with Fama-French data
Merges a xts object of returns with the ff data xts object.
}
\examples{
ffMergeXTS(y,ff_data)
}
