% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{ffModelStepLM_sub}
\alias{ffModelStepLM_sub}
\title{Generates an lm model using stepwise regression}
\usage{
ffModelStepLM_sub(y, ff_data, s = NULL, e = NULL, n = NULL)
}
\arguments{
\item{y}{Fund returns}

\item{ff_data}{Fama-French data}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of observations}
}
\value{
lm model
}
\description{
Generates an lm model using stepwise regression
}
\examples{
ffModelStepLM_sub(y,ff_data)
}
