% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{getPrices}
\alias{getPrices}
\title{Get prices for one or more symbols}
\usage{
getPrices(symbols, startDate = "1970-01-01", freq = "M",
  endDate = Sys.Date())
}
\arguments{
\item{symbols}{Tickers of the mutual funds and ETFs}

\item{startDate}{default is 1970-01-01}

\item{freq}{M for monthly or D for daily}

\item{endDate}{default is today's date (Sys.Date)}
}
\value{
xts object with prices. Index will be yearmon for monthly data
}
\description{
Get prices for one or more symbols
}
\examples{
getPrices("IVV")
getPrices("IVV", startDate="2015-12-31", freq="D")
}
