% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{getPricesAndReturns}
\alias{getPricesAndReturns}
\title{Get prices and returns for a set of symbols}
\usage{
getPricesAndReturns(symbols, startDate = "1970-01-01", endDate = Sys.Date())
}
\arguments{
\item{symbols}{List of symbols}

\item{startDate}{Start date, default is 1970-01-01}

\item{endDate}{End date, default is today's date Sys.Date()}
}
\value{
List of 3 items. Prices is an xts object with one column per symbol. returns.daily 
is a list with one item per symbol containing daily returns.  returns.monthly contains the monthly returns.
}
\description{
Get prices and returns for a set of symbols
}
\examples{
getPricesAndReturns(c("FNDB","IVV","SPY"))
}
