% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{getRiskFree}
\alias{getRiskFree}
\title{Get price and returns for 13 Week Treasury Bills}
\usage{
getRiskFree()
}
\value{
List with 4 xts objects: yield, prices, returns.daily, and returns.monthly
}
\description{
Get price and returns for 13 Week Treasury Bills
}
\examples{
getRiskFree()
}
