% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{makeCompleteMonths}
\alias{makeCompleteMonths}
\title{Makes sure the last month in an xts object is a full trading month    
This is used when converting to monthly returns. If the last price was not at the end 
of a month (e.g. 2/13/2018) a return will still be produced for that month (Feb 2018). 
We may not want to use a partial month for some calculations such as when calculating
the return for the last x months, so we eliminate those types of months.}
\usage{
makeCompleteMonths(x)
}
\arguments{
\item{x}{xts object}
}
\value{
xts object
}
\description{
Makes sure the last month in an xts object is a full trading month    
This is used when converting to monthly returns. If the last price was not at the end 
of a month (e.g. 2/13/2018) a return will still be produced for that month (Feb 2018). 
We may not want to use a partial month for some calculations such as when calculating
the return for the last x months, so we eliminate those types of months.
}
\examples{
makeCompleteMonths(x)
}
