% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{rbsa_bootstrap}
\alias{rbsa_bootstrap}
\title{Bootstrap of returns-based style analysis (RBSA)}
\usage{
rbsa_bootstrap(r.fund, r.style, n = 120L, method = "normalized",
  leverage = TRUE, width = 30, selection = "AIC", scale = 12,
  seed = NULL)
}
\arguments{
\item{r.fund}{Fund returns}

\item{r.style}{Style returns}

\item{n}{Number of trials (bootstrap)}

\item{method}{Method from Factor Analyticss package's style.fit function}

\item{leverage}{Leverage from Factor Analyticss package's style.fit function}

\item{width}{Number of observations in a window}

\item{selection}{Selection from Factor Analytics package's style.fit function}

\item{scale}{Number of periods in a year}

\item{seed}{If not NULL (default) this is used in the set.seed function.}
}
\value{
List containing: weights - xts object with one row per moving window containing the weights; 
meanSDofWeights - mean of the standard deviation of the columns of the weights.  Lower values represent 
more consistency of the weightings of the styles; regressStats - xts of the regression stats for each window 
including the rsquared (R2), tracking error (TE), mean absolute error (MAE), and root mean square error (RMSE);
fundReturn are the returns of the fund over each window; benchReturn are the returns of a benchmark defined by the 
style weight of the window (returns are annualized for periods exceeding one year); excessReturn is the fund 
return less the benchmark return.
}
\description{
Bootstrap of returns-based style analysis (RBSA)
}
\examples{
rbsa_rolling(r.fudn, r.style)

}
