% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{rbsa_calc}
\alias{rbsa_calc}
\title{Calcualate a RBSA fit and regression statistics}
\usage{
rbsa_calc(y, x, method = "normalized", leverage = TRUE, selection = "AIC",
  scale = 12, trace = 0)
}
\arguments{
\item{y}{Fund series}

\item{x}{Style serices}

\item{method}{method from Factor Analyticss package's style.fit function}

\item{leverage}{leverage from Factor Analyticss package's style.fit function}

\item{selection}{selection from Factor Analytics package's style.fit function}

\item{scale}{number of periods in a year}
}
\value{
List with weights, regStats, fund return, benchmark (style) return, excess return
}
\description{
Calcualate a RBSA fit and regression statistics
}
\examples{
rbsa_calc(y,x)

}
