% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{regressStats}
\alias{regressStats}
\title{Calculate regression statistics}
\usage{
regressStats(pred, y, scale = 12)
}
\arguments{
\item{pred}{Predicted (yhat) values}

\item{y}{Actual values}

\item{scale}{number of periods in a year (default=12)}
}
\value{
Vector with RSquared, TE (tracking error), MAE (mean absolute error),
  and RMSE (root mean square error)
}
\description{
Calculate regression statistics
}
\examples{
regressStats(pred, y)
}
