% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{scrapeQuoteSummary}
\alias{scrapeQuoteSummary}
\title{Scrape quote summary from Yahoo Finance}
\usage{
scrapeQuoteSummary(symbol)
}
\arguments{
\item{symbol}{Ticker of fund}
}
\value{
List with data from quote summary
}
\description{
Scrape quote summary from Yahoo Finance
}
\examples{
scrapeQuoteSummary("SPY")
}
