% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{ffModelStepLM}
\alias{ffModelStepLM}
\title{Generates list of lm models using step-wise regression}
\usage{
ffModelStepLM(y, ff_data, s = NULL, e = NULL, n = NULL)
}
\arguments{
\item{y}{Fund returns}

\item{ff_data}{Fama-French data}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of periods}
}
\value{
List of lm models
}
\description{
Generates list of lm models using step-wise regression
}
\examples{
ffModelStepLM(y,ff_data)
}
