% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{RBSA}
\alias{RBSA}
\title{Returns based style anal}
\usage{
RBSA(r.fund, r.style, s = NULL, e = Null, n = Null,
  method = "constrained", leverage = FALSE)
}
\arguments{
\item{r.fund}{Fund returns (xts)}

\item{r.style}{Style returns (xts)}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of observations}

\item{method}{Method from Factor Analyticss package's style.fit function}

\item{leverage}{Leverage from Factor Analyticss package's style.fit function}
}
\value{
Style weights
}
\description{
Returns based style anal
}
\examples{
RBSA(r.fund, r.style)
}
