% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{rbsa}
\alias{rbsa}
\title{Returns based style analysis}
\usage{
rbsa(r.fund, r.style, s = NULL, e = NULL, n = NULL,
  method = "normalized", leverage = TRUE, selection = "AIC", scale = 12)
}
\arguments{
\item{r.fund}{fund returns (xts)}

\item{r.style}{style returns (xts)}

\item{s}{start date}

\item{e}{end date}

\item{n}{number of observations}

\item{method}{method from Factor Analyticss package's style.fit function}

\item{leverage}{leverage from Factor Analyticss package's style.fit function}

\item{selection}{selection from Factor Analytics package's style.fit function}

\item{scale}{number of periods in a year}
}
\value{
List of 3: weights, R.squared, and adj.R.squared
}
\description{
Returns based style analysis
}
\examples{
RBSA(r.fund, r.style) 

}
