% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utilities.R
\name{RBSA_rolling}
\alias{RBSA_rolling}
\title{RBSA over a rolling window}
\usage{
RBSA_rolling(r.fund, r.style, s = NULL, e = NULL, n = NULL,
  method = "constrained", leverage = FALSE, width = 60)
}
\arguments{
\item{r.fund}{Fund returns (xts)}

\item{r.style}{Style returns (xts)}

\item{s}{Start date}

\item{e}{End date}

\item{n}{Number of Observations}

\item{method}{Method from Factor Analyticss package's style.fit function}

\item{leverage}{Leverage from Factor Analyticss package's style.fit function}

\item{width}{Number of observations in a window}
}
\value{
xts object with one row per moving window containing the weights
}
\description{
RBSA over a rolling window
}
\examples{
RBSA_rolling(r.fund, r.style)
}
